刷新题 2251

+++ LeetCode/.0001-1000/0076-min_window_substr.cpp
@@ -10,31 +10,28 @@
 输出: "BANC"
 
 说明：
-	如果 S 中不存这样的子串，则返回空字符串 ""。
-	如果 S 中存在这样的子串，我们保证它是唯一的答案。
+  如果 S 中不存这样的子串，则返回空字符串 ""。
+  如果 S 中存在这样的子串，我们保证它是唯一的答案。
 */
 
 // https://leetcode.com/problems/minimum-window-substring/discuss/26808/Here-is-a-10-line-template-that-can-solve-most-'substring'-problems
 // 抄的
 string minWindow(string s, string t)
 {
-	int M[128] = { 0 };
+	int M[128] = {0};
 	int count = static_cast<int>(t.length());
 	int len = static_cast<int>(s.length());
 	int start = 0, window = INT_MAX, h = 0, i, d;
 	for (i = 0; i < count; ++i)
-		M[t[i]] += 1;
-	for (i = 0; i < len;)
-	{
+		M[static_cast<int>(t[i])] += 1;
+	for (i = 0; i < len;) {
 		d = s[i];
 		M[d] -= 1;
 		if (M[d] >= 0)
 			--count;
 		++i;
-		if (count == 0)
-		{
-			while (count == 0)
-			{
+		if (count == 0) {
+			while (count == 0) {
 				d = s[h];
 				if (M[d] == 0)
 					++count;
@@ -42,8 +39,7 @@ string minWindow(string s, string t)
 				++h;
 			}
 			d = i - h + 1;
-			if (d < window)
-			{
+			if (d < window) {
 				window = d;
 				start = h - 1;
 			}
@@ -53,7 +49,6 @@ string minWindow(string s, string t)
 		window = 0;
 	return s.substr(start, window);
 }
-
 
 int main()
 {
