+++ LeetCode/.0001-1000/0205-isomorphic_str.cpp
@@ -28,19 +28,17 @@
 bool isIsomorphic(string s, string t)
 {
 	int len = static_cast<int>(s.length());
-	char a[256] = { 0 }, b[256] = { 0 };
-	for (int i = 0; i < len; ++i)
-	{
-		if (a[s[i]] && (a[s[i]] != t[i]))
+	char a[256] = {0}, b[256] = {0};
+	for (int i = 0; i < len; ++i) {
+		if (a[(int)s[i]] && (a[(int)s[i]] != t[i]))
 			return false;
-		if (b[t[i]] && (b[t[i]] != s[i]))
+		if (b[(int)t[i]] && (b[(int)t[i]] != s[i]))
 			return false;
-		a[s[i]] = t[i];
-		b[t[i]] = s[i];
+		a[(int)s[i]] = t[i];
+		b[(int)t[i]] = s[i];
 	}
 	return true;
 }
-
 
 int main()
 {
